criando Repository

+++ src/main/java/br/com/banco/repositories/ContaRepository.java
@@ -1,14 +1,11 @@
 package br.com.banco.repositories;
 
-
 import br.com.banco.model.Conta;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface ContaRepository extends JpaRepository<Conta, Integer> {
+public interface ContaRepository extends JpaRepository<Conta, Long> {
 
 
 }
